# ergosense-app/server/src/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ErgoSense API",
    "version": "1.0.0",
    "description": "API REST multi-tenant para SST, ergonomia, GRO, PGR, AET, psicossocial, eSocial e compliance."
  },
  "servers": [
    {
      "url": "/api",
      "description": "Gateway relativo"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/sectors": {
      "get": {
        "tags": [
          "sectors"
        ],
        "summary": "GET /api/sectors",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        }
      },
      "post": {
        "tags": [
          "sectors"
        ],
        "summary": "POST /api/sectors",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/access-requests": {
      "post": {
        "tags": [
          "access-requests"
        ],
        "summary": "POST /api/access-requests",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/public/support-contact": {
      "get": {
        "tags": [
          "Público"
        ],
        "summary": "GET /api/public/support-contact",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        }
      },
      "post": {
        "tags": [
          "Público"
        ],
        "summary": "POST /api/public/support-contact",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/collaborators": {
      "get": {
        "tags": [
          "collaborators"
        ],
        "summary": "GET /api/collaborators",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        }
      },
      "post": {
        "tags": [
          "collaborators"
        ],
        "summary": "POST /api/collaborators",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/collaborators/:id": {
      "put": {
        "tags": [
          "collaborators"
        ],
        "summary": "PUT /api/collaborators/:id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "collaborators"
        ],
        "summary": "DELETE /api/collaborators/:id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        }
      }
    },
    "/api/analyses": {
      "get": {
        "tags": [
          "analyses"
        ],
        "summary": "GET /api/analyses",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        }
      },
      "post": {
        "tags": [
          "analyses"
        ],
        "summary": "POST /api/analyses",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/analyses/:id": {
      "delete": {
        "tags": [
          "analyses"
        ],
        "summary": "DELETE /api/analyses/:id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        }
      }
    },
    "/api/analyses/:id/video": {
      "get": {
        "tags": [
          "analyses"
        ],
        "summary": "GET /api/analyses/:id/video",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        }
      }
    },
    "/api/reports": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "GET /api/reports",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "400": {
            "description": "Requisição inválida"
          },
          "401": {
            "description": "Não autenticado"
          },
          "403": {
            "description": "Sem permissão"
          },
          "429": {
            "description": "Rate limit"
          },
          "500": {
            "description": "Erro interno"
          }
        }
      }
    },
...
  }
}